Move mapping related functions from utils to mappings
