+++ src/ggrc-client/js/models/mappers/tests/mappings_spec.js
@@ -6,6 +6,8 @@
 import * as Utils from '../../../plugins/utils/models-utils';
 import Mappings from '../mappings';
 import {widgetModules} from '../../../plugins/utils/widgets-utils';
+import Permission from '../../../permission';
+import TreeViewConfig from '../../../apps/base_widgets';
 
 describe('Mappings', function () {
   let allTypes = [];
@@ -170,6 +172,97 @@ describe('Mappings', function () {
         expect(EXPECTED_GROUPS).toEqual(resultGroups);
         expect(expectedModels.sort()).toEqual(resultModels.sort());
       });
+    });
+  });
+
+  describe('getMappingTypes() method', function () {
+    it('always returns whitelisted items', function () {
+      let list = Mappings.getMappingList('DataAsset');
+      let whitelisted = ['Hello', 'World'];
+      let result = Mappings.getMappingList('DataAsset', whitelisted);
+      expect(_.difference(result, list).sort()).toEqual(whitelisted.sort());
+    });
+
+    it('always remove forbidden items', function () {
+      let forbidden = ['Policy', 'Process', 'Product', 'Program'];
+      let list = Mappings.getMappingList('DataAsset');
+      let result = Mappings.getMappingList('DataAsset', [], forbidden);
+      expect(_.difference(list, result).sort()).toEqual(forbidden.sort());
+    });
+
+    it('always leave whitelisted and remove forbidden items', function () {
+      let forbidden = ['Policy', 'Process', 'Product', 'Program'];
+      let whitelisted = ['Hello', 'World'];
+      let list = Mappings.getMappingList('DataAsset');
+      let result = Mappings.getMappingList('DataAsset',
+        whitelisted, forbidden);
+      let input = _.difference(list, result).concat(_.difference(result, list));
+      let output = forbidden.concat(whitelisted);
+      expect(input.sort()).toEqual(output.sort());
+    });
+  });
+
+  describe('isMappableType() method', function () {
+    it('returns false for AssessmentTemplate and  any type', function () {
+      let result = Mappings.isMappableType('AssessmentTemplate', 'Program');
+      expect(result).toBe(false);
+    });
+
+    it('returns true for Program and Control', function () {
+      let result = Mappings.isMappableType('Program', 'Control');
+      expect(result).toBe(true);
+    });
+  });
+
+  describe('allowedToMap() method', () => {
+    let baseWidgets;
+    beforeAll(() => {
+      baseWidgets = TreeViewConfig.attr('base_widgets_by_type');
+      TreeViewConfig.attr('base_widgets_by_type', {
+        Type1: ['Type2'],
+      });
+    });
+
+    afterAll(() => {
+      TreeViewConfig.attr('base_widgets_by_type', baseWidgets);
+    });
+
+    it('checks mapping rules', () => {
+      spyOn(Permission, 'is_allowed_for');
+      let result =
+        Mappings.allowedToMap('Issue', 'Audit', {isIssueUnmap: false});
+      expect(result).toBeFalsy();
+      expect(Permission.is_allowed_for).not.toHaveBeenCalled();
+    });
+
+    it('checks mappable types when there is no additional mapping rules',
+      () => {
+        spyOn(Permission, 'is_allowed_for');
+        let result = Mappings.allowedToMap('DataAsset', 'Assessment');
+        expect(result).toBeFalsy();
+        expect(Permission.is_allowed_for).not.toHaveBeenCalled();
+      });
+
+    it('checks permissions to update source', () => {
+      spyOn(Permission, 'is_allowed_for').and.returnValue(true);
+      let result = Mappings.allowedToMap('DataAsset', 'AccessGroup');
+      expect(result).toBeTruthy();
+      expect(Permission.is_allowed_for)
+        .toHaveBeenCalledWith('update', 'DataAsset');
+      expect(Permission.is_allowed_for.calls.count()).toEqual(1);
+    });
+
+    it('checks permissions to update target', () => {
+      spyOn(Permission, 'is_allowed_for').and.returnValue(true);
+      let source = new can.Map({type: 'DataAsset'});
+      let target = new can.Map({type: 'AccessGroup'});
+      let result = Mappings.allowedToMap(source, target);
+      expect(result).toBeTruthy();
+      expect(Permission.is_allowed_for.calls.count()).toEqual(2);
+      expect(Permission.is_allowed_for.calls.argsFor(0))
+        .toEqual(['update', source]);
+      expect(Permission.is_allowed_for.calls.argsFor(1))
+        .toEqual(['update', target]);
     });
   });
